+++ src/main/java/link/myrecipes/front/service/MemberService.java
@@ -2,11 +2,8 @@ package link.myrecipes.front.service;
 
 import link.myrecipes.front.dto.User;
 import link.myrecipes.front.dto.request.UserRequest;
-import link.myrecipes.front.dto.security.UserSecurity;
 
 public interface MemberService {
-    UserSecurity login(String username);
-
     User readMember();
 
     User createMember(UserRequest userRequest);
