+++ src/test/java/link/myrecipes/front/service/MemberServiceImplTest.java
@@ -4,7 +4,6 @@ import link.myrecipes.front.common.RestTemplateHelperImpl;
 import link.myrecipes.front.common.SecurityHelperImpl;
 import link.myrecipes.front.dto.User;
 import link.myrecipes.front.dto.request.UserRequest;
-import link.myrecipes.front.dto.security.UserSecurity;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -47,35 +46,34 @@ public class MemberServiceImplTest {
                 .build();
     }
 
-    @Test
-    public void When_로그인_정보_요청_Then_정상_반환() {
-        //given
-        UserSecurity userSecurity = UserSecurity.builder()
-                .id(1)
-                .username("user12")
-                .password("123456")
-                .accountNonExpired(true)
-                .accountNonLocked(true)
-                .credentialsNonExpired(true)
-                .enabled(true)
-                .build();
-        given(this.restTemplateHelper.getForEntity(eq(UserSecurity.class), contains("/login"))).willReturn(userSecurity);
-
-        //when
-        final UserSecurity selectedUserSecurity = this.memberService.login(userSecurity.getUsername());
-
-        //then
-        assertThat(selectedUserSecurity, instanceOf(UserSecurity.class));
-        assertThat(selectedUserSecurity.getId(), is(userSecurity.getId()));
-        assertThat(selectedUserSecurity.getUsername(), is(userSecurity.getUsername()));
-        assertThat(selectedUserSecurity.getPassword(), is(userSecurity.getPassword()));
-        assertThat(selectedUserSecurity.isAccountNonExpired(), is(userSecurity.isAccountNonExpired()));
-        assertThat(selectedUserSecurity.isAccountNonLocked(), is(userSecurity.isAccountNonLocked()));
-        assertThat(selectedUserSecurity.getId(), is(userSecurity.getId()));
-        assertThat(selectedUserSecurity.isCredentialsNonExpired(), is(userSecurity.isCredentialsNonExpired()));
-        assertThat(selectedUserSecurity.isEnabled(), is(userSecurity.isEnabled()));
-
-    }
+//    @Test
+//    public void When_로그인_정보_요청_Then_정상_반환() {
+//        //given
+//        UserSecurity userSecurity = UserSecurity.builder()
+//                .id(1)
+//                .username("user12")
+//                .password("123456")
+//                .accountNonExpired(true)
+//                .accountNonLocked(true)
+//                .credentialsNonExpired(true)
+//                .enabled(true)
+//                .build();
+//        given(this.restTemplateHelper.getForEntity(eq(UserSecurity.class), contains("/login"))).willReturn(userSecurity);
+//
+//        //when
+//        final UserSecurity selectedUserSecurity = this.memberService.login(userSecurity.getUsername());
+//
+//        //then
+//        assertThat(selectedUserSecurity, instanceOf(UserSecurity.class));
+//        assertThat(selectedUserSecurity.getId(), is(userSecurity.getId()));
+//        assertThat(selectedUserSecurity.getUsername(), is(userSecurity.getUsername()));
+//        assertThat(selectedUserSecurity.getPassword(), is(userSecurity.getPassword()));
+//        assertThat(selectedUserSecurity.isAccountNonExpired(), is(userSecurity.isAccountNonExpired()));
+//        assertThat(selectedUserSecurity.isAccountNonLocked(), is(userSecurity.isAccountNonLocked()));
+//        assertThat(selectedUserSecurity.getId(), is(userSecurity.getId()));
+//        assertThat(selectedUserSecurity.isCredentialsNonExpired(), is(userSecurity.isCredentialsNonExpired()));
+//        assertThat(selectedUserSecurity.isEnabled(), is(userSecurity.isEnabled()));
+//    }
 
     @Test
     public void When_회원정보_조회_Then_정상_반환() {
